Clean up some inappropriate choices of type in the bitcode reader. None of these are expected to fix any 64->32 bit real truncation issues.
llvm-svn: 229153

+++ llvm/include/llvm/Bitcode/BitstreamReader.h
@@ -258,8 +258,8 @@ public:
     AF_DontAutoprocessAbbrevs = 2
   };
 
-      /// Advance the current bitstream, returning the next entry in the stream.
-      BitstreamEntry advance(unsigned Flags = 0) {
+  /// Advance the current bitstream, returning the next entry in the stream.
+  BitstreamEntry advance(unsigned Flags = 0) {
     while (1) {
       unsigned Code = ReadCode();
       if (Code == bitc::END_BLOCK) {
@@ -301,7 +301,7 @@ public:
 
   /// Reset the stream to the specified bit number.
   void JumpToBit(uint64_t BitNo) {
-    uintptr_t ByteNo = uintptr_t(BitNo/8) & ~(sizeof(word_t)-1);
+    size_t ByteNo = size_t(BitNo/8) & ~(sizeof(word_t)-1);
     unsigned WordBitNo = unsigned(BitNo & (sizeof(word_t)*8-1));
     assert(canSkipToPos(ByteNo) && "Invalid location");
 
@@ -315,7 +315,7 @@ public:
   }
 
   void fillCurWord() {
-    if (Size != 0 && NextChar >= (unsigned)Size)
+    if (Size != 0 && NextChar >= Size)
       report_fatal_error("Unexpected end of file");
 
     // Read the next word from the stream.
